refactor: keep one md file per component with all framework implementations
