--- packages/storybook/src/components-twig/icon-twig.md
@@ -1,15 +1,0 @@
-<!-- @license CC0-1.0 -->
-
-# Rijkshuisstijl Community Twig icon component
-
-Dit component is een wrapper om een SVG element. Door dit component te gebruiken heeft deze constant dezelfde styling en grootte.
-
-## Gebruik
-
-Het Twig icon component kan maar op 1 manier worden gebruikt veregeleken met het [React icon component](?path=/docs/rhc-icon--docs).
-
-Standaard is er de optie om een icon property mee te geven met dit component, wat staat voor een van de ondersteunde icon IDs uit de [icon set](?path=/docs/rhc-templates-default-icon-set--docs) van de Rijkshuisstijl Community.
-
-```twig
-{% include '@rhc/icon.twig' %}
-```
